backend_python/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
autogen-agentchat==0.4.5
autogen-core==0.4.5
autogen-ext==0.4.5
azure-core==1.32.0
azure-identity==1.20.0
beautifulsoup4==4.13.3
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cobble==0.1.4
cryptography==44.0.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
isort==6.0.0
Jinja2==3.1.5
jiter==0.8.2
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
lxml==5.3.0
Mako==1.3.9
mammoth==1.9.0
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==3.0.2
more-itertools==10.6.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.2
openai==1.61.1
openapi-core==0.19.4
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
parse==1.20.2
pathable==0.4.4
pathspec==0.12.1
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
platformdirs==4.3.6
portalocker==2.10.1
prance==23.6.21.0
propcache==0.3.0
protobuf==5.29.3
puremagic==1.28
pybars4==0.9.13
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
PyMeta3==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
semantic-kernel==1.22.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.38
starlette==0.45.3
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
XlsxWriter==3.2.2
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
